[HMTM-205] - Logging update
