+++ src/common/enums.py
@@ -8,9 +8,3 @@ class Environments(str, Enum):
     LOCAL = "LOCAL"
 
 
-class Levels(Enum):
-    DEBUG = "DEBUG"
-    INFO = "INFO"
-    WARNING = "WARNING"
-    ERROR = "ERROR"
-    CRITICAL = "CRITICAL"
