+++ src/logging/enums.py
@@ -1,0 +1,9 @@
+from enum import Enum
+
+
+class Levels(Enum):
+    DEBUG = "DEBUG"
+    INFO = "INFO"
+    WARNING = "WARNING"
+    ERROR = "ERROR"
+    CRITICAL = "CRITICAL"
